actualizacion carrito restar inventario

+++ src/POS/components/ListaCompras.jsx
@@ -1,23 +1,26 @@
 import React from 'react'
 
-export const ListaCompras = ({ listaCompras, incrementItem, removeProduct, deleteItem }) => {
+export const ListaCompras = ({ detallesVenta, incrementItem, removeProduct, deleteItem }) => {
+    // Verifica si detallesVenta está vacío
+    if (detallesVenta.length === 0) {
+        return <div>No hay productos en la lista de compras.</div>;
+    }
+
     return (
         <>
             <div className="col col-sm-12 col-md-9">
                 <ul className="list-group">
-                    {listaCompras.map((item) => (
-                        <li key={item.codigo} className="list-group-item">
+                    {detallesVenta.map((item, index) => (
+                        <li key={index} className="list-group-item">
                             <div className="d-flex justify-content-between">
                                 <div>
-                                    Codigo:{item.codigo}- {item.descripcion} - Cantidad: {item.cantidad} - Subtotal: ${parseFloat(item.cantidad * item.precioSalida)
-                                    }
+                                    Codigo: {item.codigoProducto} - {item.descripcion} - Cantidad: {item.cantidad} - Subtotal: ${parseFloat(item.cantidad * item.precioUnitario)}
                                 </div>
                                 <div>
                                     <span className="btn" onClick={() => incrementItem(item)}>+</span>
                                     <span className="btn" onClick={() => removeProduct(item)}>-</span>
                                     <span className="btn" onClick={() => deleteItem(item)}>X</span>
                                 </div>
-
                             </div>
                         </li>
                     ))}
@@ -26,3 +29,4 @@ export const ListaCompras = ({ listaCompras, incrementItem, removeProduct, delet
         </>
     )
 }
+
